addressed render issue related to cleanDB.ts

+++ server/src/seeds/cleanDb.ts
@@ -3,11 +3,20 @@ import db from '../config/connection.js';
 
 export default async (modelName: "Question", collectionName: string) => {
   try {
-    let modelExists = await models[modelName].db.db.listCollections({
+    //Commented on 2/7 due to error with seeding in render
+    // let modelExists = await models[modelName].db.db.listCollections({
+    //   name: collectionName
+    // }).toArray()
+    
+    // if (!db) {
+    //   throw new Error('Error');
+    // }
+
+    const modelExists = await models[modelName]?.db?.db?.listCollections({
       name: collectionName
     }).toArray()
 
-    if (modelExists.length) {
+    if (modelExists && modelExists.length) {
       await db.dropCollection(collectionName);
     }
   } catch (err) {
